+++ SurveyPlatform.API/Controllers/UsersController.cs
@@ -2,8 +2,8 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using SurveyPlatform.API.DTOs.Requests;
-using SurveyPlatform.BLL;
 using SurveyPlatform.BLL.Models;
+using SurveyPlatform.BLL.Services;
 using SurveyPlatform.DTOs.Responses;
 
 namespace SurveyPlatform.Controllers
