+++ SurveyPlatform.BLL/Services/PollService.cs
@@ -5,12 +5,12 @@ using SurveyPlatform.BLL.Helpers;
 using SurveyPlatform.BLL.Models;
 using SurveyPlatform.DAL.Entities;
 using SurveyPlatform.DAL.Interfaces;
-namespace SurveyPlatform.BLL;
+namespace SurveyPlatform.BLL.Services;
 public class PollService(
-        IPollRepository pollRepository, 
-        IMapper mapper, 
-        IHttpContextAccessor httpContextAccessor, 
-        IOptionRepository optionRepository, 
+        IPollRepository pollRepository,
+        IMapper mapper,
+        IHttpContextAccessor httpContextAccessor,
+        IOptionRepository optionRepository,
         UserService userService
     )
 {
@@ -57,7 +57,7 @@ public class PollService(
         await pollRepository.DeletePollAsync(id);
     }
 
-    public async Task<PollModel> AddPollResponseAsync(Guid pollId,Guid optionId)
+    public async Task<PollModel> AddPollResponseAsync(Guid pollId, Guid optionId)
     {
         var httpContext = httpContextAccessor.HttpContext;
         var userId = JwtHelper.GetUserIdFromToken(httpContext);
@@ -65,17 +65,17 @@ public class PollService(
         var option = await optionRepository.GetOptionByIdAsync(optionId);
         var poll = await pollRepository.GetPollWithResponsesAsync(pollId);
 
-        if (poll == null) 
+        if (poll == null)
             throw new EntityNotFoundException($"Poll {pollId} not found");
-        if (option == null) 
+        if (option == null)
             throw new EntityNotFoundException($"Option {optionId} not found");
-        if (user == null) 
+        if (user == null)
             throw new EntityNotFoundException($"User {userId} not found");
-        if (poll.Id != option.PollId) 
+        if (poll.Id != option.PollId)
             throw new EntityConflictException($"Option ID[{optionId}] doesn't belong to Poll[{pollId}]");// Вариант ответа не из этого опроса
-        if (poll.Responses?.Any(m=>m.UserId==user.Id) == true) 
+        if (poll.Responses?.Any(m => m.UserId == user.Id) == true)
             throw new EntityConflictException($"User already responded to this poll");// Человек уже голосовал в этом опросе
-        
+
         var pollResponse = new PollResponse
         {
             PollId = pollId,
